Handling Spotify login is now at the app level
